requirements.txt
asgiref==3.4.1
Django==4.0.1
django-debug-toolbar==3.2.4
sqlparse==0.4.2
